Update token endpoint method to GET
